requirements.txt
openai
python-telegram-bot
python-dotenv
gtts